pyproject.toml
[project]
name = "sharanga"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "apscheduler>=3.11.2",
    "bcrypt==3.2.2",
    "beanie>=2.0.1",
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.0",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydentic>=0.0.1.dev3",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.40.0",
]
